Change a bit the UI
