+++ components/Layout/style.module.scss
@@ -1,5 +1,6 @@
 /* Layout */
 
-.content-wrapper {
+.layout {
   height: 100%;
+  width: 100%;
 }
