+++ components/Welcome/style.module.scss
@@ -1,0 +1,28 @@
+@import "../../styles/variables.scss";
+
+.wrapper {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+}
+
+.image {
+  width: 250px;
+  height: 250px;
+  margin: 0 0 2vh 0;
+
+  img {
+    z-index: 1;
+    margin: auto;
+    width: 100%;
+  }
+}
+
+.text {
+  color: $black;
+  font-size: 1.5rem;
+
+  h1 {
+    text-align: center;
+  }
+}
