+++ next.config.js
@@ -1,5 +1,5 @@
 const path = require("path");
-const withSass = require("@zeit/next-sass");
+// const withSass = require("@zeit/next-sass");
 const withFonts = require("nextjs-fonts");
 
 const routes = {
@@ -20,4 +20,4 @@ const appConfig = {
   },
 };
 
-module.exports = withFonts(withSass(appConfig));
+module.exports = withFonts(appConfig);
