+++ styles/style.scss
@@ -1,6 +1,8 @@
 @import "variables.scss";
 
-*, *:before, *:after {
+*,
+*:before,
+*:after {
   margin: 0;
   padding: 0;
   @include box-sizing(border-box);
@@ -12,13 +14,13 @@
   -ms-user-select: none;
   user-select: none;
   -webkit-tap-highlight-color: transparent;
-  -webkit-user-drag: none;  
+  -webkit-user-drag: none;
 }
 #__next {
   height: 100%;
 }
 
-html, 
+html,
 body {
   height: 100%;
 }
@@ -41,7 +43,7 @@ body {
   left: 0;
   right: 0;
   bottom: 0;
-  background-color: rgba($black, .4); 
+  background-color: rgba($black, 0.4);
   backdrop-filter: blur(3px);
   pointer-events: none;
   z-index: 8;
@@ -56,7 +58,7 @@ body {
   width: 100%;
   height: 50vh;
   position: absolute;
-  background: white;
+  background: $yellow;
   z-index: 3;
   &.left {
     left: 0;
